Legion Assignment Handler
